requirements.txt
###### Requirements without Version Specifiers ######
pandas
pillow
pyyaml